mktrepo5 and mktrepo6 edited

+++ src/main/java/com/aristowebapi/dto/MktRepo5.java
@@ -7,16 +7,16 @@ public interface MktRepo5 {
 	public int getSdepo_code();
 	public int getDepo_code();
 	public String getDepo_name();
-	public long getSaleval();
-	public long getTargetval();
-	public long getLysval();
+	public double getSaleval();
+	public double getTargetval();
+	public double getLysval();
 	public int getfs();
 	public String getSmname();
 	public String getEmname();
 	public String getBrname();
 	public String getGp_name();
-	public long getCrval();
-	public long getPisale();
+	public double getCrval();
+	public double getPisale();
 
 
 }
